+++ src/parser/category_parser.rs
@@ -3,9 +3,9 @@ use once_cell::sync::Lazy;
 use crate::eh_config;
 
 pub struct Category {
-    pub color: usize,
+    pub color: u32,
     pub string: String,
-    pub value: usize,
+    pub value: u32,
 }
 
 impl From<&String> for Category {
@@ -29,8 +29,8 @@ impl From<&str> for Category {
     }
 }
 
-impl From<usize> for Category {
-    fn from(value: usize) -> Self {
+impl From<u32> for Category {
+    fn from(value: u32) -> Self {
         let strings = CATEGORY_STRINGS.lock().unwrap();
         match value {
             eh_config::MISC => Category {
@@ -94,30 +94,30 @@ impl From<usize> for Category {
 
 // Use it for homepage
 const NONE: i8 = -1;
-const UNKNOWN: usize = 0x400;
+const UNKNOWN: u32 = 0x400;
 
-const ALL_CATEGORY: usize = UNKNOWN - 1;
+const ALL_CATEGORY: u32 = UNKNOWN - 1;
 
 // DOUJINSHI|MANGA|ARTIST_CG|GAME_CG|WESTERN|NON_H|IMAGE_SET|COSPLAY|ASIAN_PORN|MISC;
 
-const BG_COLOR_DOUJINSHI: usize = 0xfff44336;
-const BG_COLOR_MANGA: usize = 0xffff9800;
-const BG_COLOR_ARTIST_CG: usize = 0xfffbc02d;
-const BG_COLOR_GAME_CG: usize = 0xff4caf50;
-const BG_COLOR_WESTERN: usize = 0xff8bc34a;
-const BG_COLOR_NON_H: usize = 0xff2196f3;
-const BG_COLOR_IMAGE_SET: usize = 0xff3f51b5;
-const BG_COLOR_COSPLAY: usize = 0xff9c27b0;
-const BG_COLOR_ASIAN_PORN: usize = 0xff9575cd;
-const BG_COLOR_MISC: usize = 0xfff06292;
-const BG_COLOR_UNKNOWN: usize = 0x00000000;
+const BG_COLOR_DOUJINSHI: u32 = 0xfff44336;
+const BG_COLOR_MANGA: u32 = 0xffff9800;
+const BG_COLOR_ARTIST_CG: u32 = 0xfffbc02d;
+const BG_COLOR_GAME_CG: u32 = 0xff4caf50;
+const BG_COLOR_WESTERN: u32 = 0xff8bc34a;
+const BG_COLOR_NON_H: u32 = 0xff2196f3;
+const BG_COLOR_IMAGE_SET: u32 = 0xff3f51b5;
+const BG_COLOR_COSPLAY: u32 = 0xff9c27b0;
+const BG_COLOR_ASIAN_PORN: u32 = 0xff9575cd;
+const BG_COLOR_MISC: u32 = 0xfff06292;
+const BG_COLOR_UNKNOWN: u32 = 0x00000000;
 
 // Remove [XXX], (XXX), {XXX}, ~XXX~ stuff
 const PATTERN_TITLE_PREFIX: &str = r#"^(?:(?:\([^\)]*\))|(?:\[[^\]]*\])|(?:\{[^\}]*\})|(?:~[^~]*~)|\s+)*"#;
 // Remove [XXX], (XXX), {XXX}, ~XXX~ stuff and something like ch. 1-23
 const PATTERN_TITLE_SUFFIX: &str = r#"(?:\s+ch.[\s\d-]+)?(?:(?:\([^\)]*\))|(?:\[[^\]]*\])|(?:\{[^\}]*\})|(?:~[^~]*~)|\s+)*$"#;
 
-const CATEGORY_VALUES: [usize; 11] = [
+const CATEGORY_VALUES: [u32; 11] = [
     eh_config::MISC,
     eh_config::DOUJINSHI,
     eh_config::MANGA,
